# aceresults/docs/api/openapi.YAML
openapi: 3.0.3
info:
  title: ACE Results API
  version: 1.0.0
  description: |
    Backend contract for the ACE Engineering College Results Portal.

    **Authentication mechanisms**
    - **Student session**: JWT in the `student_session` httpOnly cookie, set by `POST /api/login`. Valid 12h. Revoked instantly when `sessionVersion` changes.
    - **Staff session (NextAuth)**: ADMIN / HOD users sign in via `/api/auth/*` (credentials provider). JWT session cookie carrying `role` and `department`.

    **Response conventions (current state)**
    Most endpoints return `{ success, message }`. Registration, login and password-change endpoints return `{ message }` only and rely on the HTTP status. See `docs/api/INVENTORY.md` for inconsistencies planned for standardisation.
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://your-domain.com
    description: Production
...
paths:
...
  /api/upload:
    post:
      tags: [HOD]
      summary: Upload and process a T-Sheet PDF
      description: |
        Parses the PDF, validates each student record, upserts `Result` documents by (rollNumber, semester), and creates Student accounts for roll numbers not yet present.
        Records failing validation are returned in `rejected`, not saved.
      security: [{ staffSession: [] }]
      x-required-role: HOD
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, semester]
              properties:
                file: { type: string, format: binary, description: T-Sheet PDF }
                semester: { type: string, example: I-II, description: "I-I … IV-II" }
                branch: { type: string, description: Defaults to the HOD's department }
                batch: { type: string }
      responses:
        "200":
          description: Processed (or no valid records, with `success:false`)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  rejectedCount: { type: integer }
                  rejected: { type: array, items: { type: object } }
        "400": { description: Missing file or semester, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { $ref: "#/components/responses/NotAuthenticated" }
        "500": { $ref: "#/components/responses/ServerError" }
...
components:
  securitySchemes:
    studentSession:
      type: apiKey
      in: cookie
      name: student_session
      description: JWT set by POST /api/login.
    staffSession:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session for ADMIN / HOD (`__Secure-next-auth.session-token` over HTTPS).
  schemas:
...
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        error: { type: string, description: Present on 500 responses }
...
  responses:
    NotAuthenticated:
      description: Not authenticated / not authorized for this role
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }